+++ evo1/atb/__init__.py
@@ -1,4 +1,4 @@
-from evo1.atb.base import SeqATBCombat
+from evo1.atb.base import ATBAction, ATBActor, ATBPlan, SeqATBCombat
 from evo1.atb.encounter import Encounter, EncounterID, calc_next_encounter
 from evo1.atb.entity import atb_stats_from_memory
 from evo1.atb.farming import FarmingGoal, SeqATBmove2D
@@ -6,6 +6,9 @@ from evo1.atb.manual import SeqATBCombatManual
 from evo1.atb.predict import predict_attack
 
 __all__ = [
+    "ATBAction",
+    "ATBActor",
+    "ATBPlan",
     "Encounter",
     "EncounterID",
     "SeqATBmove2D",